requirements.txt
# Core dependencies
torch>=2.6.0
torchvision>=0.21.0
torchaudio>=2.6.0
transformers>=4.51.0
vllm>=0.8.2
ray>=2.53.0

# Configuration
hydra-core>=1.3.2
omegaconf>=2.3.0

# Data processing
tensordict>=0.6.2
torchdata>=0.11.0
pandas>=3.0.0
numpy>=1.26.4
pyarrow>=23.0.0
datasets>=4.5.0

# Training framework
verl>=0.4.0
accelerate>=1.12.0
peft>=0.18.1
flash-attn>=2.7.4

# API & networking
openai>=2.21.0
pydantic>=2.12.0

# Evaluation
scikit-learn>=1.8.0
scipy>=1.17.0
sentence-transformers>=5.2.3

# Utilities
tqdm>=4.67.0
wandb>=0.25.0
pillow>=12.1.0
einops>=0.8.2
safetensors>=0.7.0
tokenizers>=0.21.4
